feat: problem4안의매서드  ReverseGame으로 옮겨 책임 이양. 영어가 아니면 뒤집지 않는지 확인하는 테스트 코드 추가.
